@@ -1,32 +1,27 @@
 {
-  "name": "guoshao-reader-pwa",
+  "name": "guoshao-reader",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "copy-to-clipboard": "^3.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-horizontal-scrolling-menu": "^2.1.0",
-    "react-router-cache-route": "^1.11.1",
-    "react-scripts": "4.0.3",
-    "react-scroll-to-top": "^1.0.5",
-    "web-vitals": "^1.0.1",
-    "workbox-background-sync": "^5.1.3",
-    "workbox-broadcast-update": "^5.1.3",
-    "workbox-cacheable-response": "^5.1.3",
-    "workbox-core": "^5.1.3",
-    "workbox-expiration": "^5.1.3",
-    "workbox-google-analytics": "^5.1.3",
-    "workbox-navigation-preload": "^5.1.3",
-    "workbox-precaching": "^5.1.3",
-    "workbox-range-requests": "^5.1.3",
-    "workbox-routing": "^5.1.3",
-    "workbox-strategies": "^5.1.3",
-    "workbox-streams": "^5.1.3"
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4",
+    "workbox-background-sync": "^6.4.2",
+    "workbox-broadcast-update": "^6.4.2",
+    "workbox-cacheable-response": "^6.4.2",
+    "workbox-core": "^6.4.2",
+    "workbox-expiration": "^6.4.2",
+    "workbox-google-analytics": "^6.4.2",
+    "workbox-navigation-preload": "^6.4.2",
+    "workbox-precaching": "^6.4.2",
+    "workbox-range-requests": "^6.4.2",
+    "workbox-routing": "^6.4.2",
+    "workbox-strategies": "^6.4.2",
+    "workbox-streams": "^6.4.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -53,11 +48,6 @@
     ]
   },
   "devDependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "axios": "^0.21.1",
-    "html-react-parser": "^1.2.8",
-    "react-infinite-scroll-component": "^6.1.0",
-    "react-router-dom": "^5.2.0"
+    "tailwindcss": "^3.1.8"
   }
 }
